chore: Updated example app
