+++ examples/basic-integration/app/src/main/AndroidManifest.xml
@@ -18,15 +18,15 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".SaveCardSuccessfulActivity"
+            android:name=".screens.SaveCardSuccessfulActivity"
             android:exported="false" />
 
         <activity
-            android:name=".PaymentSuccessfulActivity"
+            android:name=".screens.PaymentSuccessfulActivity"
             android:exported="false" />
 
         <activity
-            android:name=".ClickToPaySuccessfulActivity"
+            android:name=".screens.ClickToPaySuccessfulActivity"
             android:exported="false" />
 
         <activity
