+++ examples/basic-integration/app/src/main/java/com/example/demoandroid/MainActivity.kt
@@ -1,22 +1,15 @@
 package com.example.demoandroid
 
 import android.app.AlertDialog
-import android.content.DialogInterface
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.widget.Button
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import com.deuna.maven.DeunaSDK
-import com.deuna.maven.checkout.domain.*
-import com.deuna.maven.element.domain.ElementsError
-import com.deuna.maven.initCheckout
-import com.deuna.maven.initElements
-import com.deuna.maven.initPaymentWidget
-import com.deuna.maven.payment_widget.domain.PaymentWidgetCallbacks
 import com.deuna.maven.shared.*
-import com.deuna.maven.shared.domain.UserInfo
+import com.example.demoandroid.screens.PaymentSuccessfulActivity
 import org.json.JSONObject
 
 
@@ -25,15 +18,18 @@ const val DEBUG_TAG = "👀 DeunaSDK"
 
 @Suppress("UNCHECKED_CAST")
 class MainActivity : AppCompatActivity() {
-    private val deunaSdk = DeunaSDK(
+    val deunaSdk = DeunaSDK(
         environment = Environment.SANDBOX,
         publicApiKey = "YOUR_API_KEY",
     )
 
-    private val orderToken: String
+    val context: Context
+        get() = this@MainActivity
+
+    val orderToken: String
         get() = findViewById<EditText>(R.id.inputOrderToken).text.toString().trim()
 
-    private val userToken: String?
+    val userToken: String?
         get() {
             val text = findViewById<EditText>(R.id.inputUserToken).text.toString().trim()
             return text.ifEmpty { null }
@@ -52,7 +48,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         findViewById<Button>(R.id.savePaymentMethodButton).setOnClickListener {
-            saveCard()
+            initElementsVaultWidget()
         }
 
         findViewById<Button>(R.id.clickToPayButton).setOnClickListener {
@@ -60,7 +56,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun handlePaymentSuccess(data: Json) {
+    fun handlePaymentSuccess(data: Json) {
         Intent(this@MainActivity, PaymentSuccessfulActivity::class.java).apply {
             putExtra(
                 PaymentSuccessfulActivity.EXTRA_JSON_ORDER,
@@ -70,245 +66,18 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun showPaymentErrorAlertDialog(metadata: PaymentsError.Metadata) {
+    fun showPaymentErrorAlertDialog(metadata: PaymentsError.Metadata) {
         val builder: AlertDialog.Builder = AlertDialog.Builder(this)
         builder.setTitle(metadata.code)
         builder.setMessage(metadata.message)
 
         builder.setPositiveButton(
             "Aceptar",
-            DialogInterface.OnClickListener { dialog, which -> // Code to execute when OK button is clicked
-                dialog.dismiss()
-            },
-        )
+        ) { dialog, _ -> // Code to execute when OK button is clicked
+            dialog.dismiss()
+        }
 
         val alertDialog: AlertDialog = builder.create()
         alertDialog.show()
-    }
-
-    private fun showPaymentWidget() {
-        deunaSdk.initPaymentWidget(
-            context = this,
-            orderToken = orderToken,
-            cssFile = "YOUR_THEME_UUID", // optional
-            callbacks = PaymentWidgetCallbacks().apply {
-                onSuccess = { data ->
-                    deunaSdk.close()
-                    handlePaymentSuccess(data)
-                }
-                onCanceled = {
-                    Log.d(DEBUG_TAG, "Payment was canceled by user")
-                }
-                onCardBinDetected = { cardBinMetadata, refetchOrder ->
-                    Log.d(DEBUG_TAG, "cardBinMetadata: $cardBinMetadata")
-                    if (cardBinMetadata != null) {
-
-                        deunaSdk.setCustomStyle(
-                            data = JSONObject(
-                                """
-                        {
-                          "theme": {
-                            "colors": {
-                              "primaryTextColor": "#023047",
-                              "backgroundSecondary": "#8ECAE6",
-                              "backgroundPrimary": "#F2F2F2",
-                              "buttonPrimaryFill": "#FFB703",
-                              "buttonPrimaryHover": "#FFB703",
-                              "buttonPrimaryText": "#000000",
-                              "buttonPrimaryActive": "#FFB703"
-                            }
-                          },
-                          "HeaderPattern": {
-                            "overrides": {
-                              "Logo": {
-                                "props": {
-                                  "url": "https://images-staging.getduna.com/ema/fc78ef09-ffc7-4d04-aec3-4c2a2023b336/test2.png"
-                                }
-                              }
-                            }
-                          }
-                        }
-                        """
-                            ).toMap()
-                        )
-
-                        refetchOrder { order ->
-                            Log.d(DEBUG_TAG, "onCardBinDetected > refetchOrder: $order")
-                        }
-
-                    }
-                }
-                onInstallmentSelected = { metadata, refetchOrder ->
-                    Log.d(DEBUG_TAG, "installmentMetadata: $metadata")
-                    refetchOrder { order ->
-                        Log.d(DEBUG_TAG, "onInstallmentSelected > refetchOrder: $order")
-                    }
-                }
-                onClosed = {
-                    Log.d(DEBUG_TAG, "Widget was closed")
-                }
-                onError = { error ->
-                    Log.e(DEBUG_TAG, "Error type: ${error.type}, metadata: ${error.metadata}")
-                    when (error.type) {
-                        PaymentsError.Type.INITIALIZATION_FAILED,
-                        PaymentsError.Type.NO_INTERNET_CONNECTION -> {
-                            deunaSdk.close()
-                            if (error.metadata != null) {
-                                showPaymentErrorAlertDialog(error.metadata!!)
-                            }
-                        }
-
-                        else -> {}
-                    }
-                }
-                onEventDispatch = { type, data ->
-                    Log.d(DEBUG_TAG, "onEventDispatch ${type.name}: $data")
-                }
-            },
-            userToken = userToken,
-            paymentMethods = listOf(
-                mapOf(
-                    "payment_method" to "voucher",
-                    "processors" to listOf("daviplata", "nequi_push_voucher")
-                ),
-                mapOf(
-                    "payment_method" to "paypal",
-                    "processors" to listOf("paypal")
-                )
-            ),
-            checkoutModules = listOf(
-                mapOf("name" to "module_name")
-            )
-        )
-    }
-
-
-    private fun startPaymentProcess() {
-        deunaSdk.initCheckout(
-            context = this,
-            orderToken = orderToken,
-            callbacks = CheckoutCallbacks().apply {
-                onSuccess = { data ->
-                    Log.d(DEBUG_TAG, "Payment success $data")
-                    deunaSdk.close()
-                    handlePaymentSuccess(data)
-                }
-                onError = { error ->
-                    Log.e(DEBUG_TAG, "Error type: ${error.type}, metadata: ${error.metadata}")
-                    when (error.type) {
-                        PaymentsError.Type.PAYMENT_ERROR,
-                        PaymentsError.Type.ORDER_COULD_NOT_BE_RETRIEVED,
-                        PaymentsError.Type.INITIALIZATION_FAILED -> {
-                            deunaSdk.close()
-                            if (error.metadata != null) {
-                                showPaymentErrorAlertDialog(error.metadata!!)
-                            }
-                        }
-
-                        else -> {}
-                    }
-                }
-                onCanceled = {
-                    Log.d(DEBUG_TAG, "Payment was canceled by user")
-                }
-                onEventDispatch = { type, data ->
-                    Log.d(DEBUG_TAG, "onEventDispatch ${type.name}: $data")
-                    when (type) {
-                        CheckoutEvent.changeAddress, CheckoutEvent.changeCart -> {
-                            deunaSdk.close()
-                        }
-
-                        else -> {}
-                    }
-                }
-                onClosed = {
-                    Log.d(DEBUG_TAG, "Widget was closed")
-                }
-            },
-            userToken = userToken,
-        )
-    }
-
-
-    private fun saveCard() {
-        deunaSdk.initElements(
-            context = this,
-            userToken = userToken,
-            userInfo = if (userToken == null) UserInfo(
-                firstName = "Darwin",
-                lastName = "Morocho",
-                email = "dmorocho@deuna.com",
-            ) else null,
-            callbacks = ElementsCallbacks().apply {
-                onSuccess = { data ->
-                    val metadata = (data["metadata"] as Json)["createdCard"] as Json
-                    deunaSdk.close()
-                    Intent(this@MainActivity, SaveCardSuccessfulActivity::class.java).apply {
-                        putExtra(
-                            SaveCardSuccessfulActivity.EXTRA_CREATED_CARD,
-                            JSONObject(metadata).toString()
-                        )
-                        startActivity(this)
-                    }
-                }
-                onEventDispatch = { type, data ->
-                    Log.d(DEBUG_TAG, "onEventDispatch ${type.name}: $data")
-                }
-                onError = {
-                    Log.e(ERROR_TAG, it.type.message)
-                    deunaSdk.close()
-                }
-                onCanceled = {
-                    Log.d(DEBUG_TAG, "Saving card was canceled by user")
-                }
-                onClosed = {
-                    Log.d(DEBUG_TAG, "Widget was closed")
-                }
-            },
-        )
-    }
-
-
-    /**
-     * Show the widget that processes a payment using Click to Pay.
-     */
-    private fun clickToPay() {
-        deunaSdk.initElements(
-            context = this,
-            userInfo = UserInfo(
-                // required for click_to_pay
-                firstName = "Darwin",
-                lastName = "Morocho",
-                email = "dmorocho@deuna.com",
-            ),
-            types = listOf(
-                mapOf(
-                    "name" to ElementsWidget.CLICK_TO_PAY
-                )
-            ),
-            callbacks = ElementsCallbacks().apply {
-                onSuccess = {
-                    deunaSdk.close()
-                    Intent(this@MainActivity, ClickToPaySuccessfulActivity::class.java).apply {
-                        startActivity(this)
-                    }
-                }
-                onEventDispatch = { type, data ->
-                    Log.d(DEBUG_TAG, "onEventDispatch ${type.name}: $data")
-                }
-                onError = { error ->
-                    Log.e(ERROR_TAG, error.type.message)
-                    Log.e(ERROR_TAG, error.metadata?.code ?: "")
-                    Log.e(ERROR_TAG, error.metadata?.message ?: "")
-
-                    if (error.type == ElementsError.Type.INITIALIZATION_FAILED) {
-                        deunaSdk.close()
-                    }
-                }
-                onClosed = {
-                    Log.d(DEBUG_TAG, "Widget was closed")
-                }
-            },
-        )
     }
 }
