+++ examples/basic-integration/app/src/main/java/com/example/demoandroid/screens/ClickToPaySuccessfulActivity.kt
@@ -1,4 +1,4 @@
-package com.example.demoandroid
+package com.example.demoandroid.screens
 
 import android.annotation.SuppressLint
 import android.content.Intent
@@ -6,6 +6,8 @@ import android.os.Bundle
 import android.view.View
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import com.example.demoandroid.MainActivity
+import com.example.demoandroid.R
 
 class ClickToPaySuccessfulActivity : AppCompatActivity() {
