+++ examples/basic-integration/app/src/main/res/layout/activity_thank_you.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SaveCardSuccessfulActivity">
+    tools:context=".screens.SaveCardSuccessfulActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
